backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
edge-tts==7.2.7
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.128.0
flatbuffers==25.12.19
frozenlist==1.8.0
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
langchain==1.2.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.2.6
openai==2.15.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
websockets==15.0.1
Werkzeug==3.1.5
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
